@@ -4,9 +4,12 @@
   "description": "Preggies back end service.",
   "main": "app.js",
   "scripts": {
-    "start": "nodemon src/server.js",
+    "start": "node lib/server.js",
+    "dev": "nodemon --delay 20ms --exec NODE_ENV=development ts-node -- src/server.ts",
     "test": "jest",
     "lint": "eslint src --fix",
+    "clean:build": "rm -rf lib/*",
+    "build": "npm run clean:build && tsc",
     "pretest": "npm run lint"
   },
   "author": "Oyemade",
@@ -29,18 +32,19 @@
     ]
   },
   "dependencies": {
+    "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
+    "express-mongo-sanitize": "^1.3.2",
+    "helmet": "^3.16.0",
     "mongoose": "^5.8.7",
     "morgan": "^1.9.1",
-    "xss-clean": "^0.1.1",
-    "express-mongo-sanitize": "^1.3.2",
-    "cors": "^2.8.5",
-    "helmet": "^3.16.0"
+    "nanoexpress-pro": "1.11.1",
+    "xss-clean": "^0.1.1"
   },
   "devDependencies": {
+    "@types/node": "13.11.0",
     "eslint": "^5.16.0",
-    "eslint-config-airbnb": "^17.1.0",
     "eslint-config-prettier": "^4.1.0",
     "eslint-plugin-import": "^2.17.2",
     "eslint-plugin-jest": "23.8.2",
@@ -48,9 +52,12 @@
     "jest": "25.2.7",
     "nodemon": "2.0.2",
     "parcel-bundler": "^1.12.3",
-    "prettier": "^1.17.0"
+    "prettier": "^1.17.0",
+    "ts-jest": "25.3.1",
+    "ts-node": "8.8.2",
+    "typescript": "3.8.3"
   },
   "engines": {
-    "node": "^10"
+    "node": "^12"
   }
 }
